Simplify build setup and test runner
Node gained capability to run TS files directly, so we don't have to
pre-compile them anymore. This changes the test runner to run TS files
directly. Because of that, the build script could be significantly
simplified—the test build mode was removed and generation of the banner
was used to add the UserScript banner.