Use a map instead of an object for unparsed stats
That fits better because stats aren't exactly a fixed set of properties.
While they don't change often, they still change sometimes so it'd be
pointless to maintain a fixed-shape object when we only need a handful
of them. Moreover, because of the lack of the fixed shape, an object may
have worse performance because of many different shapes it can take at
runtime, causing runtimes to deoptimize the internal shape structure.