Detect a game client version only once
Previously, the client detection was run inside functions defined in the
`#external` module. This is unnecessary because the addon won't be able
to support dynamic changes anyway. (It won't be able to detect when a
change happens to reapply necessary monkey-patches to client functions.)
It is expected that a page reload is performed when a switch is made.

Additionally, for now, this throws an error when run in an unsupported
environment. In the future, this should be changed to something safer so
that it doesn't lead to unhandled exceptions and errors in the console
on arbitrary websites. Alternatively, the `@match` directives in the
user script banner might be changed to include only known game domains,
but that will be an unnecessary maintenance burden, since those domains
sometimes get added or removed.