Count only stats that an item explicitly contains
Previously, the addon was attempting to invoke counters for all known
stats even when an item didn't have a particular statistic. This is
obviously unnecessary, but it was an unfortunate consequence of the
stats representation.

By changing from a fixed-shape object to a map, we can iterate only over
existing stats. This creates much smaller objects at runtime and makes
the code more straightforward. (Plus there are much less console logs,
so devtools run smoother.)

Similarly to the commit with composable bonus counter, this also removes
tests temporarily so that I don't have to spend time fixing them while
there's still potential for massive API changes.