Add counters for armor stats
These are the first stats that depend on the rarity modifier, so several
bugs with handling of that modifier were discovered. They all stem from
the same thing: stats don't have individual rarity modifiers, but rather
it's a single modifier bound to an item that can affect multiple stats.

Therefore, this changes the decomposition from including the modifier
explicitly, to including a boolean that only designates whether a stat
is influenced by the modifier.