Sort out the terminology (part 1)
This got to the point that even I have some trouble keeping up with what
means what. Some terms are used interchangeably even though they should
not, so it's time to sort them out.

Starting out with renames of files and directories, the module `#count`
was renamed to `#decompose` which is an accurate description of the
actual process happening there. Previously, “count” was overloaded and
now it'll refer to the count of bonuses specifically.