Bonuses shouldn't have default values in `StatDecompositionState`
Defaulting bonuses makes it impossible to differentiate between unknown
values and actually counted values that happen to match the default ones.

This also updates the `rarityDependent` counter because this change
indicated that its current behavior is incorrect. Instead of assuming
that the state will always contain a bonus count, it should use the
remaining stat value for control flow. That also significantly
simplifies that counter.