Use an object instead of a map for storing stat formatters
This is mainly for consistency with other parts of the codebase. When a
mapping isn't dynamic, that is, it's initialized once and no properties
are added or removed later, then object literal syntax is more pleasant.

Performance-wise this should not matter. If anything, I'd expect objects
to be more performant, but that's not important enough for me to bother
checking that. (At least now.)